+++ app/server/src/controllers/configController.ts
@@ -1,24 +1,18 @@
 import { Request, Response } from "express";
 import { ConfigReader } from "../configReader";
 import { ErrorCode, jsonResponseError, jsonResponseSuccess } from "../jsonResponse";
+import { AppLocals } from "../types";
 
 export class ConfigController {
-    private readonly _configReader: ConfigReader;
-
-    private readonly _appsPath: string;
-
-    constructor(configReader: ConfigReader, appsPath: string) {
-        this._configReader = configReader;
-        this._appsPath = appsPath;
-    }
-
-    private _readAppConfigFile = (appName: string) => {
-        return this._configReader.readConfigFile(this._appsPath, `${appName}.config.json`);
+    private static _readAppConfigFile = (appName: string, appsPath: string, configReader: ConfigReader) => {
+        return configReader.readConfigFile(appsPath, `${appName}.config.json`);
     };
 
-    getConfig = (req: Request, res: Response) => {
+    static getConfig = (req: Request, res: Response) => {
         const { appName } = req.params;
-        const config = this._readAppConfigFile(appName);
+        const { configReader, appsPath } = req.app.locals as AppLocals;
+
+        const config = this._readAppConfigFile(appName, appsPath, configReader);
         if (config) {
             jsonResponseSuccess(config, res);
         } else {
