+++ app/server/src/routes/apps.ts
@@ -1,0 +1,7 @@
+import { AppsController } from "../controllers/appsController";
+
+const router = require("express").Router();
+
+router.get("/:appName", AppsController.getApp);
+
+module.exports = router;
