+++ app/server/src/routes/config.ts
@@ -1,17 +1,7 @@
 import { ConfigController } from "../controllers/configController";
-import {ConfigReader} from "../configReader";
 
-const express = require("express");
+const router = require("express").Router();
 
-const app = express();
-const router = express.Router();
+router.get("/:appName", ConfigController.getConfig);
 
-console.log("CONFIG PATH: "  + app.get("configPath")); //THIS ISN'T GETTING PICKED UP -  NOT USED?
-console.log("APPS LOCAL TEST:" + app.locals.testy)
-
-const configReader = new ConfigReader(app.get("configPath"));
-const configController = new ConfigController(configReader, app.get("appsPath"));
-
-app.use("config", router);
-router.get(`/:appName`, configController.getConfig);
 module.exports = router;
