+++ app/server/src/routes/index.ts
@@ -1,6 +1,18 @@
-const router = require('express').Router();
+import { Application, Request, Response } from "express";
+import { IndexController } from "../controllers/indexController";
 
-router.use('/odin', require('./odin'));
-router.use('/config', require('./config'));
+module.exports = (app: Application) => {
+    const router = require("express").Router();
 
-module.exports = router;
+    router.get("/", IndexController.getIndex);
+    router.use("/odin", require("./odin"));
+    router.use("/config", require("./config"));
+    router.use(`/${app.locals.appsPath}`, require("./apps"));
+
+    router.use((req: Request, res: Response) => {
+        const { url } = req;
+        res.status(404).render("page-not-found", { url });
+    });
+
+    return router;
+};
