+++ app/server/src/routes/odin.ts
@@ -1,8 +1,8 @@
 import { OdinController } from "../controllers/odinController";
 
-const router = require('express').Router();
+const router = require("express").Router();
 
-router.get(`/runner`, OdinController.getRunner);
-router.get(`/model`, OdinController.getModel);
+router.get("/runner", OdinController.getRunner);
+router.get("/model", OdinController.getModel);
 
 module.exports = router;
