+++ app/server/src/types.ts
@@ -1,4 +1,12 @@
+import { ConfigReader } from "./configReader";
+
 export interface WodinConfig {
     port: number,
     appsPath: string
+}
+
+export interface AppLocals {
+    appsPath: string,
+    configPath: string,
+    configReader: ConfigReader
 }
